Move Equality related stuff to own package
There will probably be more equality related classes in the future.

This simply makes it clearer what these relate to
