+++ scalameta/contrib/src/main/scala/scala/meta/contrib/equality/Syntactically.scala
@@ -1,4 +1,4 @@
-package scala.meta.contrib
+package scala.meta.contrib.equality
 
 import scala.language.implicitConversions
 import scala.meta.Tree
