+++ scalameta/contrib/src/test/scala/scala/meta/contrib/EqualSuite.scala
@@ -1,8 +1,9 @@
 package scala.meta
 package contrib
 import scala.meta.testkit._
-
 import org.scalatest.FunSuite
+
+import scala.meta.contrib.equality.{Structurally, Syntactically}
 
 class EqualSuite extends FunSuite {
   val a: Stat = "val x = 2 // foo".parse[Stat].get
